% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proteusTMT.R
\docType{data}
\name{evi}
\alias{evi}
\title{Evidence data}
\format{A data frame with 601,792 rows and 19 columns}
\description{
Evidence data from a TMT proteomics experiment in 5 conditions, 3 replicates each.
}
