% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proteusTMT.R
\docType{data}
\name{meta}
\alias{meta}
\title{Proteomics metadata}
\format{A \code{data.frame} with 30 rows and 5 columns (experiment, measure, sample, replicate).}
\description{
Metadata describing a TMT proteomics experiment in 5 conditions with 3 replicates each.
}
