% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proteusTMT.R
\docType{data}
\name{proteusTMT}
\alias{proteusTMT}
\title{Proteus example TMT data}
\format{Four objects from different stages of data processing:
\describe{
  \item{\code{evi}}{Pre-processed evidence data (data frame)}
  \item{\code{meta}}{Metadata (data frame)}
  \item{\code{pepdat}}{Peptide data (\code{proteusData} object)}
  \item{\code{prodat}}{Protein data (\code{proteusData} object)}
}
This package also contains a formatted metadata file (see examples below).}
\description{
Example data set for \code{proteus} package. It comes from a TMT MS/MS experiment in
ten time-points and three replicates each.
See \href{http://www.mcponline.org/content/17/6/1184}{Crozier at al. 2018},
\href{https://www.ebi.ac.uk/pride/archive/projects/PXD008741}{PRIDE PXD008741}.
}
\examples{
library(proteusTMT)
data(proteusTMT)
meta
head(pepdat$tab)
head(prodat$tab)

metadataFile <- system.file("extdata", "metadata.txt", package="proteusTMT")
}
