% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proteusTMT.R
\docType{data}
\name{proteusTMT}
\alias{proteusTMT}
\title{Proteus example TMT data}
\format{Four objects from different stages of data processing:
\describe{
  \item{\code{evi}}{Pre-processed evidence data (data frame)}
  \item{\code{meta}}{Metadata (data frame)}
  \item{\code{pepdat}}{Peptide data (\code{proteusData} object)}
  \item{\code{prodat}}{Protein data (\code{proteusData} object)}
}
This package also contains the full original evidence file and formatted metadata file (see examples below).}
\description{
Example data set for \code{proteus} package. It comes from a TMT MS/MS experiment in two conditions
and three replicates each. See https://www.ebi.ac.uk/pride/archive/projects/PXD005968.
}
\examples{
library(proteusTMT)
data(proteusTMT)
meta
head(pepdat$tab)
head(prodat$tab)

evidenceFile <- system.file("extdata", "evidence.txt.gz", package="proteusTMT")
metadataFile <- system.file("extdata", "metadata.txt", package="proteusTMT")
}
